chore: foram feitas refatorações no código e adicionado alguns novos components
