+++ projects/crypto-currency/iluminados/app/src/components/Login/styles.js
@@ -1,0 +1,7 @@
+import styled from 'styled-components'
+
+export const DivContainerBtn = styled.div`
+  &&{
+    align-self: start;
+  }
+`
