+++ projects/crypto-currency/iluminados/app/src/components/Register/styles.js
@@ -1,0 +1,7 @@
+import styled from "styled-components"
+
+export const DivContainerBtn = styled.div`
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+`
